First pass of cleaning Excel_Engine
