+++ Excel_Engine/Excel_Engine.csproj
@@ -65,21 +65,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Compute\Contents.cs" />
-    <Compile Include="Compute\Note.cs" />
-    <Compile Include="Convert\FromExcel\FromExcelFormula.cs" />
-    <Compile Include="Convert\ToExcel\ToExcelFormula.cs" />
-    <Compile Include="Query\Selection.cs" />
-    <Compile Include="Query\Filename.cs" />
-    <Compile Include="Query\Formula.cs" />
-    <Compile Include="Compute\NumberFormat.cs" />
     <Compile Include="Create\Reference.cs" />
-    <Compile Include="Convert\FromExcel\FromExcel.cs" />
-    <Compile Include="Convert\ToExcel\ToExcel.cs" />
+    <Compile Include="Convert\FromExcel.cs" />
+    <Compile Include="Convert\ToExcel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Query\CleanList.cs" />
-    <Compile Include="Query\RefText.cs" />
-    <Compile Include="Query\Caller.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -90,6 +79,10 @@
       <Name>Excel_oM</Name>
       <Private>False</Private>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Compute\" />
+    <Folder Include="Query\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
