--- Excel_Engine/Query/Filename.cs
@@ -1,57 +1,0 @@
-/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *
- *
- * The BHoM is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3.0 of the License, or
- * (at your option) any later version.
- *
- * The BHoM is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.
- */
-
-using BH.oM.Excel;
-using BH.oM.Reflection.Attributes;
-using System.ComponentModel;
-using System.Text.RegularExpressions;
-
-namespace BH.Engine.Excel
-{
-    public static partial class Query
-    {
-        /*******************************************/
-        /**** Methods                           ****/
-        /*******************************************/
-
-
-        [Description("Gets the filename of this workbook.")]
-        [Output("A filename.")]
-        public static string Filename()
-        {
-            return Caller().Filename();
-        }
-
-        /*******************************************/
-
-        [Description("Gets the filename of the workbook of the given reference.")]
-        [Input("reference", "A reference for which to find the filename.")]
-        [Output("A filename.")]
-        public static string Filename(this Reference reference)
-        {
-            string reftext = reference.RefText();
-            return Regex.Match(reftext, @"\[(.*)\]").Groups[1].Value;
-        }
-
-        /*******************************************/
-    }
-}
